# docs/openapi/public.yaml
openapi: 3.1.0
info:
  title: City Flood MVP API
  version: 0.1.0
paths:
...
  /v1/telemetry/scenarios:
    get:
      tags:
      - telemetry
      summary: Telemetry Scenarios
      operationId: telemetry_scenarios_v1_telemetry_scenarios_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ScenarioRunOut'
                type: array
                title: Response Telemetry Scenarios V1 Telemetry Scenarios Get
              example:
              - scenario_id: scenario_123
                name: heavy_rain_high_river
                start_ts: '2026-02-01T00:00:00Z'
                end_ts: '2026-02-02T00:00:00Z'
        '400':
          $ref: '#/components/responses/InvalidArgument'
  /v1/telemetry/qa:
    get:
      tags:
      - telemetry
      summary: Telemetry Qa
      operationId: telemetry_qa_v1_telemetry_qa_get
      parameters:
      - name: city_id
        in: query
        required: true
        schema:
          type: string
          title: City Id
      - name: day
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: YYYY-MM-DD
          title: Day
        description: YYYY-MM-DD
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryQaOut'
                title: Response Telemetry Qa V1 Telemetry Qa Get
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /v1/assets/{asset_id}/observations:
    get:
      tags:
      - telemetry
      summary: Observations
      operationId: observations_v1_assets__asset_id__observations_get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: metric
        in: query
        required: true
        schema:
          type: string
          title: Metric
      - name: from
        in: query
        required: true
        schema:
          type: string
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          title: To
      - name: granularity
        in: query
        required: false
        schema:
          type: string
          description: Timescale time_bucket interval, e.g. 1m,5m,1h
          default: 5m
          title: Granularity
        description: Timescale time_bucket interval, e.g. 1m,5m,1h
      - name: agg
        in: query
        required: false
        schema:
          type: string
          description: avg|min|max
          default: avg
          title: Agg
        description: avg|min|max
      - name: include_gaps
        in: query
        required: false
        schema:
          type: boolean
          description: include empty buckets as null values
          default: false
          title: Include Gaps
        description: include empty buckets as null values
      - name: scenario_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: scenario id or 'latest'
          title: Scenario Id
        description: scenario id or 'latest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationsOut'
              example:
                metric: fill_ratio
                series:
                - ts: '2026-02-01T10:00:00Z'
                  value: 0.42
                - ts: '2026-02-01T10:05:00Z'
                  value: 0.55
                request_id: req_example
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  schemas:
...
    ObservationOut:
      properties:
        ts:
          type: string
          format: date-time
          title: Ts
        value:
          anyOf:
          - type: number
          - type: 'null'
          title: Value
      type: object
      required:
      - ts
      - value
      title: ObservationOut
    ObservationsOut:
      properties:
        metric:
          type: string
          title: Metric
        series:
          items:
            $ref: '#/components/schemas/ObservationOut'
          type: array
          title: Series
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - metric
      - series
      - request_id
      title: ObservationsOut
    ScenarioRunOut:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        name:
          type: string
          title: Name
        start_ts:
          type: string
          format: date-time
          title: Start Ts
        end_ts:
          type: string
          format: date-time
          title: End Ts
      type: object
      required:
      - scenario_id
      - name
      - start_ts
      - end_ts
      title: ScenarioRunOut
    TelemetryQaOut:
      properties:
        city_id:
          type: string
          title: City Id
        day:
          type: string
          format: date
          title: Day
        metric:
          type: string
          title: Metric
        assets:
          type: integer
          title: Assets
        buckets_expected:
          type: integer
          title: Buckets Expected
        buckets_present:
          type: integer
          title: Buckets Present
        gaps:
          type: integer
          title: Gaps
        suspect_count:
          type: integer
          title: Suspect Count
      type: object
      required:
      - city_id
      - day
      - metric
      - assets
      - buckets_expected
      - buckets_present
      - gaps
      - suspect_count
      title: TelemetryQaOut
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
          required:
          - code
          - message
          - details
          - request_id
      required:
      - error
  responses:
    InvalidArgument:
      description: 'INVALID_ARGUMENT: Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INVALID_ARGUMENT
              message: Validation error
              details: {}
              request_id: req_example
...
    NotFound:
      description: 'NOT_FOUND: Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              details: {}
              request_id: req_example
...
    UnprocessableEntity:
      description: 'UNPROCESSABLE_ENTITY: Unprocessable entity'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNPROCESSABLE_ENTITY
              message: Unprocessable entity
              details: {}
              request_id: req_example
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security: []